refactor. ResponseHandler, Parser 클래스 분리
비대해진 RequestHandler의 메서드들을 용도에 맞게 분리함
