+++ src/main/java/webserver/WebServer.java
@@ -26,7 +26,7 @@ public class WebServer {
         try (ServerSocket listenSocket = new ServerSocket(port)) {
             logger.info("Web Application Server started {} port.", port);
 
-            ExecutorService executorService = Executors.newFixedThreadPool(16);
+//            ExecutorService executorService = Executors.newFixedThreadPool(16);
 
             // 클라이언트가 연결될때까지 대기한다.
             Socket connection;
@@ -34,10 +34,6 @@ public class WebServer {
                 CompletableFuture<Void> thread = CompletableFuture.runAsync(new RequestHandler(connection));
                 thread.join();
 
-
-//                executorService.submit(() -> new RequestHandler(connection).run())
-//                Thread thread = new Thread(new RequestHandler(connection));
-//                thread.start();
             }
         }
     }
